+++ flowserv/model/workflow/manager.py
@@ -13,7 +13,6 @@ underlying database.
 
 import git
 import os
-import shutil
 import tempfile
 
 from contextlib import contextmanager
@@ -311,6 +310,7 @@ def clone(source, repository=None):
             repository = WorkflowRepository()
         repourl, manifestpath = repository.get(source)
         sourcedir = tempfile.mkdtemp()
+        print('cloning into {}'.format(sourcedir))
         if manifestpath is not None:
             manifestpath = os.path.join(sourcedir, manifestpath)
         try:
@@ -320,4 +320,6 @@ def clone(source, repository=None):
             raise ex
         finally:
             # Make sure to cleanup by removing the created teporary folder.
-            shutil.rmtree(sourcedir)
+            # Avoid permission errors for read-only files on Windows based on
+            # https://stackoverflow.com/questions/58878089
+            git.rmtree(sourcedir)